render tiles at fractional scale
set_draw_scale/rescale_tileset take a double and round tile sizes to
whole pixels, lifting zoom off the 0.25-4x power-of-two steps.

Also pin the focused tile to the window centre: the view origin is
whole-tile quantized, so at fractional tile sizes the centred tile would
otherwise drift by up to half a tile.

+++ src/sdltiles.cpp
@@ -4102,7 +4102,7 @@ bool gamepad_available()
     return gamepad::get_controller() != nullptr;
 }
 
-void rescale_tileset( int size )
+void rescale_tileset( double size )
 {
     // zoom is calculated as powers of 2 so need to convert swap zoom between 4 and 64
     if( size <= pow( 2, get_option<int>( "SWAP_ZOOM" ) + 1 ) && use_far_tiles ) {
