+++ src/sdltiles.h
@@ -47,7 +47,7 @@ extern int fontwidth;
 // may be displayed. Actually, this is supposed to be called from init.cpp,
 // and only from there.
 void load_tileset();
-void rescale_tileset( int size );
+void rescale_tileset( double size );
 bool save_screenshot( const std::string &file_path );
 void toggle_fullscreen_window();
